requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio @ file:///croot/anyio_1745334642479/work
attrs==25.3.0
certifi==2025.6.15
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi @ file:///croot/fastapi_1724789151288/work
fastapi-cli==0.0.7
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna @ file:///croot/idna_1714398848350/work
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
outcome==1.3.0.post0
psycopg2-binary==2.9.10
pydantic @ file:///croot/pydantic_1750768261732/work
pydantic_core @ file:///croot/pydantic-core_1750754736071/work
Pygments==2.19.2
PySocks==1.7.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.8
selenium==4.34.0
shellingham==1.5.4
sniffio @ file:///croot/sniffio_1705431295498/work
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.38.6
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection @ file:///croot/typing-inspection_1746023470701/work
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
